requirements.txt
streamlit
torch
transformers
ddgs
groq
python-dotenv
numpy